% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{all_package_deps}
\alias{all_package_deps}
\title{Retrieve dependency information for all packages available on CRAN and on
Bioconductor.}
\usage{
all_package_deps(pkg_info)
}
\arguments{
\item{pkg_info}{Package information retrieved by \code{\link[=all_package_info]{all_package_info()}}.}
}
\value{
A named list representing all dependencies of each package.
}
\description{
Retrieve dependency information for all packages available on CRAN and on
Bioconductor.
}
