% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{all_package_info}
\alias{all_package_info}
\title{Retrieve package information for all packages on CRAN and Bioconductor.}
\usage{
all_package_info()
}
\value{
A \code{pkg_info} object.
}
\description{
Retrieve package information for all packages on CRAN and Bioconductor.
}
