% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{all_packages}
\alias{all_packages}
\title{Build a package tree including all packages on CRAN and Bioconductor.}
\usage{
all_packages()
}
\value{
A \code{pkg_tree} object.
}
\description{
Build a package tree including all packages on CRAN and Bioconductor.
}
