% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{as.character.pkg_tree}
\alias{as.character.pkg_tree}
\title{Cast a package tree into a vector of package names.}
\usage{
\method{as.character}{pkg_tree}(x, ...)
}
\arguments{
\item{x}{A \code{pkg_tree} object.}

\item{...}{Ignored.}
}
\value{
A character vector of package names.
}
\description{
Cast a package tree into a vector of package names.
}
