% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{bioc_packages}
\alias{bioc_packages}
\title{Build a package tree including all Bioconductor packages}
\usage{
bioc_packages()
}
\value{
A \code{pkg_tree} object.
}
\description{
Build a package tree including all Bioconductor packages
}
