% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{build_schedule}
\alias{build_schedule}
\title{Computes a build schedule respecting inter-package dependencies.}
\usage{
build_schedule(deps)
}
\arguments{
\item{deps}{A list representing packages dependencies.}
}
\value{
Vector of packages names in build order.
}
\description{
Computes a build schedule respecting inter-package dependencies.
}
