% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{buildable_pkgs}
\alias{buildable_pkgs}
\title{List packages that can be build now.}
\usage{
buildable_pkgs(build_sched)
}
\arguments{
\item{build_sched}{A build schedule.}
}
\value{
A vector of package names.
}
\description{
List packages that can be build now.
}
