% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel_packages}
\alias{channel_packages}
\title{List packages available from a channel.}
\usage{
channel_packages(channel)
}
\arguments{
\item{channel}{A \code{conda_channel} object.}
}
\value{
A data.frame containing package versions and build numbers, with
package names used as row names.
}
\description{
List packages available from a channel.
}
