% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_build}
\alias{conda_build}
\title{Build an R package with Conda.}
\usage{
conda_build(
  pkg,
  tree,
  build_num = 0,
  dry_run = FALSE,
  variants = TRUE,
  log_dir = getwd()
)
}
\arguments{
\item{pkg}{Package name.}

\item{tree}{A \code{pkg_tree} object.}

\item{build_num}{An integer representing the build version of the package.}

\item{dry_run}{Enable dry run when \code{TRUE}.}

\item{variants}{Enable multi‑arch variant generation when \code{TRUE}.}

\item{log_dir}{Write logs to specified directory, defaulting to current
directory.}
}
\description{
This function works on the assumption that all dependencies are already
available.
}
\details{
When \code{dry-run} is \code{TRUE}, the function generates the recipe file but does not
initiate the build process. Instead, it returns the path to the temporary
directory containing the recipe.
}
