% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{conda_build_num}
\alias{conda_build_num}
\title{Lookup the build number of a package.}
\usage{
conda_build_num(pkg, channel)
}
\arguments{
\item{pkg}{The package name.}

\item{channel}{A \code{conda_channel} object.}
}
\value{
The build version as an integer if the package exists, or \code{NA}.
}
\description{
Lookup the build number of a package.
}
