% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{conda_channel}
\alias{conda_channel}
\title{Read the manifest of a Conda channel.}
\usage{
conda_channel(path)
}
\arguments{
\item{path}{Root directory of the channel}
}
\value{
A \code{conda_channel} object.
}
\description{
Read the manifest of a Conda channel.
}
