% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_build}
\alias{demo_build}
\title{Run a demonstration of the system, building the graphite package and its
dependencies.}
\usage{
demo_build()
}
\description{
Run a demonstration of the system, building the graphite package and its
dependencies.
}
