% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{lookup_custom}
\alias{lookup_custom}
\title{Retrieve custom build details for specific packages.}
\usage{
lookup_custom(pkg)
}
\arguments{
\item{pkg}{A length-one vector containing a package name.}
}
\value{
A list with the following entries, any of which could be \code{NULL}:
\item{compiler}{Package-specific compiler.}
\item{build_deps}{Build dependencies.}
\item{run_deps}{Runtime dependencies.}
\item{script}{Build script.}
}
\description{
Retrieve custom build details for specific packages.
}
