% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdeps.R
\name{lookup_sysdeps}
\alias{lookup_sysdeps}
\title{Lookup the system dependencies of a package.}
\usage{
lookup_sysdeps(pkg)
}
\arguments{
\item{pkg}{A vector of package names.}
}
\value{
A vector of system dependencies.
}
\description{
Lookup the system dependencies of a package.
}
