% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{mark_built_pkgs}
\alias{mark_built_pkgs}
\title{Record freshly built packages.}
\usage{
mark_built_pkgs(build_sched, pkgs)
}
\arguments{
\item{build_sched}{A build schedule.}

\item{pkgs}{A vector of package names.}
}
\value{
An updated version of the build schedule.
}
\description{
Record freshly built packages.
}
