% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{mark_failed_pkgs}
\alias{mark_failed_pkgs}
\title{Record build failures.}
\usage{
mark_failed_pkgs(build_sched, pkgs)
}
\arguments{
\item{build_sched}{A build schedule.}

\item{pkgs}{A vector of package names.}
}
\value{
An updated version of the build schedule.
}
\description{
Record build failures.
}
