% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{mark_pkgs_up_to_date}
\alias{mark_pkgs_up_to_date}
\title{Mark packages already built at their most recent version.}
\usage{
mark_pkgs_up_to_date(build_sched, pkgs)
}
\arguments{
\item{build_sched}{A build schedule.}

\item{pkgs}{A vector of package names.}
}
\value{
An updated version of the build schedule.
}
\description{
Mark packages already built at their most recent version.
}
