% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarch.R
\name{microarch_detect}
\alias{microarch_detect}
\title{Detect the microarchitecture of the host}
\usage{
microarch_detect()
}
\value{
The integer representing the microarchitecture level.
}
\description{
Detect the microarchitecture of the host
}
