% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{mtfrm.pkg_tree}
\alias{mtfrm.pkg_tree}
\title{Trasform a tree object into a vector of package names for matching.}
\usage{
\method{mtfrm}{pkg_tree}(x)
}
\arguments{
\item{x}{A \code{pkg_tree} object.}
}
\value{
A character vector of package names.
}
\description{
Trasform a tree object into a vector of package names for matching.
}
