% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{names.pkg_info}
\alias{names.pkg_info}
\title{Extract package names.}
\usage{
\method{names}{pkg_info}(x)
}
\arguments{
\item{x}{A \code{pkg_info} object.}
}
\value{
A vector of package names.
}
\description{
Extract package names.
}
