% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_deps}
\alias{pkg_deps}
\title{Retrieve the dependencies of a package.}
\usage{
pkg_deps(pkg_info, pkg)
}
\arguments{
\item{pkg_info}{A \code{pkg_info} object.}

\item{pkg}{A length-one vector containing  a package name.}
}
\value{
A vector of dependency names.
}
\description{
Retrieve the dependencies of a package.
}
