% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_info}
\alias{pkg_info}
\title{Retrieve package information.}
\usage{
pkg_info(tree, pkg, download = FALSE)
}
\arguments{
\item{tree}{A \code{pkg_tree} object.}

\item{pkg}{A length-one vector containing a package name.}

\item{download}{Download package tarball to extract full metadata.}
}
\value{
A list including all available package information.
}
\description{
Retrieve package information.
}
