% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_repos}
\alias{pkg_repos}
\title{List package repositories.}
\usage{
pkg_repos(tree, pkg)
}
\arguments{
\item{tree}{A \code{pkg_tree} object.}

\item{pkg}{A vector of package names.}
}
\value{
A vector of repositories corresponding to the provided packages.
}
\description{
List package repositories.
}
