% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_versions}
\alias{pkg_versions}
\title{List package versions.}
\usage{
pkg_versions(tree, pkg)
}
\arguments{
\item{tree}{A \code{pkg_tree} object.}

\item{pkg}{A vector of package names.}
}
\value{
A vector of package versions.
}
\description{
List package versions.
}
