% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{print.build_sched}
\alias{print.build_sched}
\title{Print method for \code{build_sched} objects.}
\usage{
\method{print}{build_sched}(x, ...)
}
\arguments{
\item{x}{A \code{build_sched} object.}

\item{...}{Other arguments, ignored.}
}
\description{
Print method for \code{build_sched} objects.
}
