% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{print.conda_channel}
\alias{print.conda_channel}
\title{Print method for \code{conda_channel} objects.}
\usage{
\method{print}{conda_channel}(x, ...)
}
\arguments{
\item{x}{A \code{conda_channel} object.}

\item{...}{Other arguments, ignored.}
}
\description{
Print method for \code{conda_channel} objects.
}
