% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{print.pkg_info}
\alias{print.pkg_info}
\title{Print function for \code{pkg_info} objects.}
\usage{
\method{print}{pkg_info}(x, ...)
}
\arguments{
\item{x}{A \code{pkg_info} object.}

\item{...}{Other arguments, ignored.}
}
\description{
Print function for \code{pkg_info} objects.
}
