% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{print.pkg_tree}
\alias{print.pkg_tree}
\title{Print function for \code{pkg_tree} objects.}
\usage{
\method{print}{pkg_tree}(x, ...)
}
\arguments{
\item{x}{A \code{pkg_tree} object.}

\item{...}{Other arguments, ignored.}
}
\description{
Print function for \code{pkg_tree} objects.
}
