% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{qualified_names}
\alias{qualified_names}
\title{Transform R package names in the format required by Conda.}
\usage{
qualified_names(pkgs, tree)
}
\arguments{
\item{pkgs}{A vector of package names.}

\item{tree}{A \code{pkg_tree} object.}
}
\value{
A vector of transformed package names.
}
\description{
Transform R package names in the format required by Conda.
}
