% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ronda-package.R
\docType{package}
\name{ronda-package}
\alias{ronda}
\alias{ronda-package}
\title{ronda: Build driver for Conda packages}
\description{
Automates the process of packaging R projects for distribution using the Conda package manager.
}
\author{
\strong{Maintainer}: Gabriele Sales \email{gabriele.sales@unipd.it} (\href{https://orcid.org/0000-0003-2078-5661}{ORCID})

}
\keyword{internal}
