% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{set_built_pkgs}
\alias{set_built_pkgs}
\title{Set packages as built.}
\usage{
set_built_pkgs(build_sched, pkgs)
}
\arguments{
\item{build_sched}{A build schedule.}

\item{pkgs}{A vector of package names.}
}
\value{
An updated version of the build schedule.
}
\description{
Set packages as built.
}
