% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{setdiff.pkg_tree}
\alias{setdiff.pkg_tree}
\title{Remove some packages from a package tree, preserving dependencies.}
\usage{
\method{setdiff}{pkg_tree}(x, y, ...)
}
\arguments{
\item{x}{A \code{pkg_tree} object.}

\item{y}{A vector of package names to be removed.}

\item{...}{Other arguments, ignored.}
}
\value{
Another \code{pkg_tree}.
}
\description{
This function will remove from the tree the required packages,
unless this breaks the dependency contraints. As a consequence,
the packages actually removed might just be a subset of those
required by the caller.
}
