% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{subset.pkg_info}
\alias{subset.pkg_info}
\title{Subset package information, keeping track of all required dependencies.}
\usage{
\method{subset}{pkg_info}(x, subset, ...)
}
\arguments{
\item{x}{A \code{pkg_info} object.}

\item{subset}{A vector of package names to be selected.}

\item{...}{Other arguments, ignored.}
}
\value{
Another pkg_info, including required packages and their dependencies.
}
\description{
Subset package information, keeping track of all required dependencies.
}
