% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{subset.pkg_tree}
\alias{subset.pkg_tree}
\title{Subset a package tree, keeping track of all required dependencies.}
\usage{
\method{subset}{pkg_tree}(x, subset, ...)
}
\arguments{
\item{x}{A \code{pkg_tree} object.}

\item{subset}{A vector of package names to be selected.}

\item{...}{Other arguments, ignored.}
}
\value{
Another \code{pkg_tree}, including required packages and their
dependencies.
}
\description{
Subset a package tree, keeping track of all required dependencies.
}
