% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{build_schedule}
\alias{build_schedule}
\title{Computes a build schedule respecting inter-package dependencies.}
\usage{
build_schedule(pkg_info)
}
\arguments{
\item{pkg_info}{Packages to build.}
}
\value{
A \code{build_sched} object.
}
\description{
Computes a build schedule respecting inter-package dependencies.
}
