% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{build_schedule}
\alias{build_schedule}
\title{Computes a build schedule respecting inter-package dependencies.}
\usage{
build_schedule(tree)
}
\arguments{
\item{tree}{Packages to build.}
}
\value{
A \code{build_sched} object.
}
\description{
Computes a build schedule respecting inter-package dependencies.
}
