% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_build}
\alias{conda_build}
\title{Build an R package with Conda.}
\usage{
conda_build(pkg, pkg_info)
}
\arguments{
\item{pkg}{Package name.}

\item{pkg_info}{A \code{pkg_info} object.}
}
\description{
This function works on the assumption tha all dependencies are already
available.
}
