% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_build}
\alias{conda_build}
\title{Build an R package with Conda.}
\usage{
conda_build(pkg, tree)
}
\arguments{
\item{pkg}{Package name.}

\item{tree}{A \code{pkg_tree} object.}
}
\description{
This function works on the assumption that all dependencies are already
available.
}
