% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{ronda_build}
\alias{ronda_build}
\title{Build a set of packages, including all of their dependencies.}
\usage{
ronda_build(pkgs)
}
\arguments{
\item{pkgs}{A vector of package names.}
}
\description{
Packages already available in the local Conda channel are skipped.
}
