% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{ronda_build}
\alias{ronda_build}
\title{Build a set of packages, including all of their dependencies.}
\usage{
ronda_build(pkgs, log_dir = getwd())
}
\arguments{
\item{pkgs}{A vector of package names, or a \code{pkg_tree} object.}

\item{log_dir}{Write logs to specified directory, defaulting to current
directory.}
}
\description{
Packages already available in the local Conda channel are skipped.
}
